Conflicts:
	filetypes.xml
	ignore.xml
	perl5.xml

+++ code/collectOnly.py
@@ -1,12 +1,24 @@
 #!/usr/bin/python3
 
+import gffutils
+import gffutils.gffwriter as gffwriter
 import os
+import subprocess
 import sys
+import tempfile
 from Bio import SeqIO
 
 count_sequences = 0
 length_cluster = 0
 
+#==========================================================================================================
+# COMMANDS LIST
+
+GFFREAD_W = 'gffread -g %s -W -w %s %s'
+
+GFFREAD_M = 'gffread -M -F -o %s %s'
+
+#==========================================================================================================
 
 
 def parse_only(threshold_float, wd, verbose):
@@ -18,7 +30,7 @@ def parse_only(threshold_float, wd, verbose):
         sys.stderr.write('Executing: Parse assembled consensus and EVM\n')
     for root, dirs, _ in os.walk(wd):
         for direc in dirs:
-            if 'output' in direc:
+            if direc.endswith('output'):
                 outputDir = wd + direc + '/'
                 if outputDir:
                     parse_contigs(outputDir, threshold_float, verbose)
@@ -50,11 +62,13 @@ def parse_contigs(output_assembly, threshold_float, verbose):
     fname = output_assembly + 'contig_member'
     fname_exists = os.path.isfile(fname)
     if fname_exists:
+        location_list = output_assembly.split("/")[-2].split(".")[0].split("_")[:-1]
+        location = "_".join(location_list)
         # part all the files in the tmp assembly folder
-        contigInfo = open(output_assembly + 'contig_member', 'r')
+        contig_info = open(output_assembly + 'contig_member', 'r')
         contigs = {}
         total_reads = 0
-        for line in contigInfo:
+        for line in contig_info:
             line = line.strip()
             line = line.split('\t')
             # count the reads for each assembled Unitig
@@ -66,7 +80,7 @@ def parse_contigs(output_assembly, threshold_float, verbose):
                 contigs[line[0]] = [read_number]
             # sum all the reads whitin one cluster
             total_reads += read_number
-        contigInfo.close()
+        contig_info.close()
         # calculate the number of reads considering the threshold
         threshold = total_reads * float(threshold_float)
         global count_sequences
@@ -76,27 +90,28 @@ def parse_contigs(output_assembly, threshold_float, verbose):
             # to retrieve only supported assembly
             if len(element) == 2:
                 if element[0] >= threshold and 'evm' in element[1]:
-                    real_contigs[key] = element[1] + ' ' + str(
+                    real_contigs[key] = element[1] + '_' + location + ' ' + str(
                         element[0]) + '_above_threshold_' + str(threshold) + ' loc_' + str(count_sequences)
                 elif element[0] < threshold and 'evm' in element[1]:
-                    real_contigs[key] = element[1] + ' ' + str(
+                    real_contigs[key] = element[1] + '_' + location + ' ' + str(
                         element[0]) + '_below_threshold_' + str(threshold) + ' loc_' + str(count_sequences)
             elif len(element) == 1:
                 if element[0] >= threshold:
-                    real_contigs[key] = 'Unitig' + str(count_sequences) + '_' + str(count_unitigs) + ' ' + str(
+                    real_contigs[key] = 'Unitig' + str(count_sequences) + '_' + str(count_unitigs) + '_' + location + ' ' + str(
                         element[0]) + '_above_threshold_' + str(threshold) + ' loc_' + str(count_sequences)
                     count_unitigs += 1
         # writes the outputs
-        fileAssembly = output_assembly + 'unigene_seq.new.fasta'
-        contigSeq = open(fileAssembly, 'r')
-        contigDict = SeqIO.to_dict(SeqIO.parse(contigSeq, 'fasta'))
-        output_filename = output_assembly[:-1] + '_assembled.fasta'
-        outputFile = open(output_filename, 'w')
-        for iden, write_iden in list(real_contigs.items()):
-            if iden in contigDict:
-                outputFile.write('>' + write_iden + '\n' + str(contigDict[iden].seq) + '\n')
-        contigSeq.close()
-        outputFile.close()
+        file_assembly = output_assembly + 'unigene_seq.new.fasta'
+        if os.path.exists(file_assembly):
+            contig_seq = open(file_assembly, 'r')
+            contig_dict = SeqIO.to_dict(SeqIO.parse(contig_seq, 'fasta'))
+            output_filename = output_assembly[:-1] + '_assembled.fasta'
+            output_file = open(output_filename, 'w')
+            for iden, write_iden in list(real_contigs.items()):
+                if iden in contig_dict:
+                    output_file.write('>' + write_iden + '\n' + str(contig_dict[iden].seq) + '\n')
+            contig_seq.close()
+            output_file.close()
     return
 
 
@@ -106,9 +121,8 @@ def cat_assembled(wd):
     """
     sys.stdout.write('\t###GENERATE FASTA FILE FROM CONTIGS###\n')
     wd_tmp = wd
-    fileName = wd_tmp + 'assembly.fasta_tmp1'
+    fileName = wd_tmp + 'assembly.fasta'
     testFasta = open(fileName, 'w')
-
     for root, dirs, files in os.walk(wd_tmp):
         for name in files:
             wd_fasta = os.path.join(root, name)
@@ -122,7 +136,7 @@ def cat_assembled(wd):
                         if evm[0] in above:
                             SeqIO.write(fasta_dict[evm[0]], testFasta, "fasta")
                         else:
-                            fasta_dict[above[0]].id = fasta_dict[evm[0]].id
+                            SeqIO.write(fasta_dict[evm[0]], testFasta, "fasta")
                             SeqIO.write(fasta_dict[above[0]], testFasta, "fasta")
                     elif len(evm) > 1:
                         SeqIO.write(fasta_dict[evm[0]], testFasta, "fasta")
@@ -133,8 +147,6 @@ def cat_assembled(wd):
                     SeqIO.write(fasta_dict[evm[0]], testFasta, "fasta")
                 elif len(above) > 0:
                     SeqIO.write(fasta_dict[above[0]], testFasta, "fasta")
-
-
     return fileName
 
 
@@ -152,51 +164,67 @@ def cat_assembled_all(wd):
             wd_dir = os.path.join(root, name)
             if 'fasta_output' in wd_dir:
                 wd_fasta = os.path.join(wd_dir, "unigene_seq.new.fasta")
-                fasta_sequences = SeqIO.parse(open(wd_fasta),'fasta')
-                for fasta in fasta_sequences:
-                    fasta_name = ">assembled_" + str(count_seq_ass)
-                    count_seq_ass += 1
-                    cdna = fasta_name + "\n" + str(fasta.seq) + "\n"
-                    out_file.write(cdna)
+                if os.path.exists(wd_fasta):
+                    fasta_sequences = SeqIO.parse(open(wd_fasta),'fasta')
+                    for fasta in fasta_sequences:
+                        fasta_name = ">assembled_" + str(count_seq_ass)
+                        count_seq_ass += 1
+                        cdna = fasta_name + "\n" + str(fasta.seq) + "\n"
+                        out_file.write(cdna)
     return fileName
 
 
-def add_EVM(gffread_fasta_file, tmp_assembly, merged_fasta_filename):
+def add_EVM(final_update, wd, consensus_mapped_gff3):
 
     """
-    this module looks for genes that were not used in the consensus stage. usually are gene models without long reads
-    support
     """
-    sys.stdout.write('\t###APPEND EVM NOT USED FROM CONTIGS BUILDING###\n')
-    '''Adds the EVM records that are not present in the final contig evidence'''
-    whole_fasta = open(gffread_fasta_file, 'r')
-    out_fasta_file = open(tmp_assembly, 'r')
-    outputMerged = open(merged_fasta_filename, 'w')
 
-    wholeDict = SeqIO.to_dict(SeqIO.parse(whole_fasta, 'fasta'))
-    count = 0
-    dictOut = {}
-    outFasta = SeqIO.parse(out_fasta_file, 'fasta')
-    for record in outFasta:
-        if record.id in dictOut:
-            dictOut[str(record.id) + '_' + str(count)] = str(record.seq)
-            count += 1
-        else:
-            dictOut[record.id] = str(record.seq)
-    for key in list(wholeDict.keys()):
-        if 'evm' in key and key not in dictOut:
-            ident = '>Gene' + str(count) + '_' + key
-            outputMerged.write(
-                ident + '\n' + str(wholeDict[key].seq) + '\n')
-            count += 1
-    for key, element in list(dictOut.items()):
-        ident = '>Gene' + str(count) + '_' + key
-        outputMerged.write(ident + '\n' + str(element) + '\n')
-        count += 1
+    db_evm = gffutils.create_db(final_update, ':memory:', merge_strategy='create_unique', keep_order=True)
+    ids_evm = [gene.attributes["ID"][0] for gene in db_evm.features_of_type("mRNA")]
 
-    whole_fasta.close()
-    outFasta.close()
-    outputMerged.close()
+    db_gmap = gffutils.create_db(consensus_mapped_gff3, ':memory:', merge_strategy='create_unique', keep_order=True)
+    ids_gmap_full = [gene.attributes["ID"][0] for gene in db_gmap.features_of_type("gene")]
+    ids_gmap = [gene.attributes["ID"][0].split("_")[0] for gene in db_gmap.features_of_type("gene")]
+
+    uniq_evm = [evm for evm in ids_evm if not evm in ids_gmap]
+
+    mRNA = []
+    for evm in uniq_evm:
+        for line in db_evm.parents(evm, order_by='start'):
+            mRNA.append(line.attributes["ID"][0])
+    mRNA_uniq = list(set(mRNA))
+    outfile = tempfile.NamedTemporaryFile(delete=False, prefix="additional.1.", suffix=".gff3", dir=wd)
+    gff_out_s = gffwriter.GFFWriter(outfile.name)
+
+    for name in mRNA_uniq:
+        for i in db_evm.children(name, order_by='start'):
+            gff_out_s.write_rec(i)
+        gff_out_s.write_rec(db_evm[name])
+    for name in ids_gmap_full:
+        for i in db_gmap.children(name, order_by='start'):
+            gff_out_s.write_rec(i)
+        gff_out_s.write_rec(db_gmap[name])
+    gff_out_s.close()
+
+    return outfile.name
+    #return outfile_out.name
+
+
+def transform_func(x):
+
+     if 'locus' in x.featuretype:
+         x.featuretype = "gene"
+         x.attributes['ID'] = x.attributes['locus']
+         x.source = "LoReAn"
+         return x
+     elif 'mRNA' in x.featuretype:
+         x.attributes['Parent'] = x.attributes['locus']
+         x.source = "LoReAn"
+         return x
+     else:
+         x.source = "LoReAn"
+         return x
+
 
 if __name__ == '__main__':
-    cat_assembled(*sys.argv[1:])
+    add_EVM(*sys.argv[1:])
